Various fixes and improvements.

+++ xmrstak/backend/cpu/crypto/hash.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <stdint.h>
+
 typedef unsigned char BitSequence;
-typedef unsigned long long DataLength;
+typedef uint32_t DataLength;
